finishing major refactor

+++ app/src/main/java/devs/mrp/coolyourturkey/comun/FileTimeGetter.java
@@ -4,7 +4,7 @@ import android.net.Uri;
 
 public interface FileTimeGetter {
 
-    public long fromFileDaysAgo(int daysAgo, Uri uri);
-    public long fromFileLastDays(int lastDays, Uri uri);
+    public long fromFileDaysAgo(long daysAgo, Uri uri);
+    public long fromFileLastDays(long lastDays, Uri uri);
 
 }
